requirements.txt
streamlit
langchain
langchain-community
langchain-text-splitters
langchain-google-genai
google-generativeai
python-docx
chromadb
